+++ src/main/java/argent_matter/gcyr/common/data/GCyRRecipes.java
@@ -1,6 +1,7 @@
 package argent_matter.gcyr.common.data;
 
 import argent_matter.gcyr.data.recipe.*;
+import argent_matter.gcyr.data.recipe.chemistry.ChemistryRecipes;
 import argent_matter.gcyr.data.recipe.chemistry.PolymerRecipes;
 import net.minecraft.data.recipes.FinishedRecipe;
 
@@ -12,6 +13,7 @@ public class GCyRRecipes {
         OxygenSpreaderRecipeLoader.init(provider);
         DysonSphereRecipeLoader.init(provider);
         PolymerRecipes.init(provider);
+        ChemistryRecipes.init(provider);
         MiscRecipes.init(provider);
         RocketFuelRecipes.init(provider);